ApplyElementTransformVisitor: top level class; rename to avoid clash.
